requirements.txt
django
djangorestframework 
djangorestframework-simplejwt 
celery 
redis 
requests 
mysqlclient 
python-dotenv